Rewrite tests.helpers.AsyncBufferReader
We need to intersperse reads and writes to really test the edge
cases of the connection code.  The current implementation wasn't
tight enough to do this reliably.  I dumped the hokey max-size
tracking and implemented separate tracking of read and write
positions instead.  This simplified the code a lot more than I
expected.  I also added a "someone is waiting" future that is used
to deliver frames while blocking the test code.